StrideUp/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-gis>=1.0
djangorestframework-simplejwt>=5.3.0
djoser>=2.2.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
django-leaflet>=0.29.0